+++ src/main/java/task2/sol2/MapperSol2.java
@@ -1,5 +1,6 @@
 package task2.sol2;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -9,14 +10,14 @@ import java.io.IOException;
 /**
  * Created by fotis on 08/02/16.
  */
-public class MapperSol2 extends Mapper<LongWritable, Text, LongWritable, LongWritable> {
+public class MapperSol2 extends Mapper<LongWritable, Text, LongWritable, IntWritable> {
 
 
     public void map(LongWritable key, Text value, Context context) throws IOException,InterruptedException {
         String[] array = value.toString().split("\\t");
-        int articleId = Integer.parseInt(array[0]);
+        long articleId = Long.parseLong(array[0]);
         int modifications =Integer.parseInt(array[1]);
-        context.write(new LongWritable(articleId), new LongWritable(modifications));
+        context.write(new LongWritable(articleId), new IntWritable(modifications));
     }
