+++ src/main/java/task2/sol2/ReducerSol2.java
@@ -1,5 +1,6 @@
 package task2.sol2;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 import task2.CustomPairMods;
@@ -11,7 +12,7 @@ import java.util.PriorityQueue;
 /**
  * Created by fotis on 01/02/16.
  */
-public class ReducerSol2 extends Reducer<LongWritable, LongWritable, LongWritable, LongWritable> {
+public class ReducerSol2 extends Reducer<LongWritable, IntWritable, LongWritable, IntWritable> {
     private PriorityQueue<CustomPairMods> pq = new PriorityQueue<CustomPairMods>();
     private int k;
 
@@ -20,14 +21,14 @@ public class ReducerSol2 extends Reducer<LongWritable, LongWritable, LongWritabl
         k = Integer.parseInt(context.getConfiguration().get("k"));
     }
 
-    public void reduce(LongWritable key, Iterable<LongWritable> values, Context context) throws IOException,InterruptedException {
+    public void reduce(LongWritable key, Iterable<IntWritable> values, Context context) throws IOException,InterruptedException {
         int counter = 0; //More than one value ?
         long articleId = key.get();
         CustomPairMods customPairMods = null;
 
         if (values.iterator().hasNext()){
             counter+=1;
-            long modifications = values.iterator().next().get();
+            int modifications = values.iterator().next().get();
             if (counter == 2){
                 throw new IOException();
             }
@@ -44,7 +45,7 @@ public class ReducerSol2 extends Reducer<LongWritable, LongWritable, LongWritabl
         CustomPairMods customPairMods = null;
         while (pq.size() >0){
             customPairMods = pq.poll();
-            context.write(new LongWritable(customPairMods.getArticleId()), new LongWritable(customPairMods.getModifications()));
+            context.write(new LongWritable(customPairMods.getArticleId()), new IntWritable(customPairMods.getModifications()));
         }
     }
 }
